tv-streamer/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
google-cloud-storage==2.18.2
Werkzeug==3.0.3
cloud-sql-python-connector[pg8000]==1.4.2
gunicorn==23.0.0
